franky updates some eclipse porject files *add some valgrind options to the sconstruct *add some extra error checking flags to the compile *list all the projects that need to be parsed *add some profiling code *remove the cv code within the input library (decoupling) *fix some warnings *remove kinect code from within the renderer (decoupling) *some code cleanup *remove the devil code *added some gl etxture loading code *add some code to obtain the texture from vram *add some code to save the texture to disk *addition of a camera test *improve the render object test

+++ lib/Renderer/ResourceManager.cpp
@@ -34,7 +34,7 @@ void ResourceManager::ReleaseResource( const Resource *resource )
       if (it->second == resource)
         {
           --it->second->mRefCount;
-          GCLAssert(resource->mRefCount>=0); //if we get a ref count of -1, something went really wrong;
+          GCLAssert((ssize_t)(resource->mRefCount)>=0); //if we get a ref count of -1, something went really wrong;
           return;
         }
     }
